// apps/frontend/src/api/models/phaseType.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Breakout Deliberation OS API
 * OpenAPI spec version: 1.0.0
 */

export type PhaseType = (typeof PhaseType)[keyof typeof PhaseType];

export const PhaseType = {
	video: "video",
	discussion: "discussion",
	voting: "voting",
	survey: "survey",
} as const;