// apps/frontend/src/api/models/surveyPhaseFeatureFlags.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Breakout Deliberation OS API
 * OpenAPI spec version: 1.0.0
 */

export interface SurveyPhaseFeatureFlags {
	participantCanProposeTransition?: boolean;
	/** @minimum 0 */
	transitionMinDurationSec?: number;
	/**
	 * @minimum 0
	 * @maximum 1
	 */
	transitionThreshold?: number;
	/** @exclusiveMinimum 0 */
	transitionVoteDurationSec?: number;
}